requirements.txt
openai
pypdf
Pillow